Updated to filter by the new upload.project_phase_id field when applicable. Unused code removed.
git-svn-id: https://coder.topcoder.com/tcs/clients/cronos/components@212206 fb370eea-3af6-4597-97f7-f7400a59c12a

+++ online_review_deliverables/trunk/src/java/main/com/cronos/onlinereview/deliverables/ApprovalDeliverableChecker.java
@@ -76,7 +76,7 @@ public class ApprovalDeliverableChecker extends SingleQuerySqlDeliverableChecker
     protected void fillInQueryParameters(Deliverable deliverable, PreparedStatement statement)
             throws SQLException, DeliverableCheckingException {
         statement.setLong(1, deliverable.getResource());
-        statement.setLong(2, deliverable.getSubmission().longValue());
+        statement.setLong(2, deliverable.getSubmission());
         statement.setLong(3, deliverable.getPhase());
     }
 }
